Remove client history code.
In preparation for removal of the notion of match (which in turn will
be in preparation for a new, more general way of handling parameters),
remove ClientHistory, ClientHistoryBook, HistoryWriter and related code.
These classes aren't useful and general enough to justify modifying them
to work without matches.  csv.DictWriter can be used instead of
HistoryWriter.