Reorganize code for arbitrary parameters.
Did some reorganization of server.frame and GameControl so that they
don't look at the contents of the parameter object for any required
fields.  This allows structure of the parameter object to be completely
up to the experiment designer.  To get the required parameters
previously assumed to exist in the parameter object, the server calls
a handful of get methods of the controller, which return default
values and can optionally be overridden by the derived controller class.
Also, the controller's runRound() is expected to return a boolean value
that specifies whether to run any further rounds.